+++ src/Justimmo/Exception/AuthenticationException.php
@@ -1,7 +1,7 @@
 <?php
 namespace Justimmo\Exception;
 
-class AuthenticationException extends \Exception
+class AuthenticationException extends \Exception implements JustimmoException
 {
 
 }