+++ src/Justimmo/Exception/JustimmoException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Justimmo\Exception;
+
+/**
+ * interface to catch all justimmo PHP-SDK related exceptions
+ *
+ * @package Justimmo\Exception
+ */
+interface JustimmoException
+{
+
+}