+++ src/Justimmo/Exception/MethodNotFoundException.php
@@ -1,7 +1,7 @@
 <?php
 namespace Justimmo\Exception;
 
-class MethodNotFoundException extends \Exception
+class MethodNotFoundException extends \Exception implements JustimmoException
 {
 
 }