+++ src/Justimmo/Exception/NotFoundException.php
@@ -1,7 +1,7 @@
 <?php
 namespace Justimmo\Exception;
 
-class NotFoundException extends \Exception
+class NotFoundException extends \Exception implements JustimmoException
 {
 
 }