+++ src/Justimmo/Exception/StatusCodeException.php
@@ -1,7 +1,7 @@
 <?php
 namespace Justimmo\Exception;
 
-class StatusCodeException extends \Exception
+class StatusCodeException extends \Exception implements JustimmoException
 {
 
 }